mm/compaction: respect compaction order when updating defer counter
It doesn't make sense that we reset defer counter
in compaction_defer_reset() when compaction request under the order of
compact_order_failed succeed. Fix it.

On the other hand, when deferring compaction for an order lower
than the current compact_order_failed, we can assume the lower order
will recover more quickly, so we should reset the progress made
previously on compact_defer_shift with the higher order."

[vbabka@suse.cz: suggest better commit description]
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>