mm/compaction: skip useless pfn when updating cached pfn
Cached pfn is used to determine the start position of scanner
at next compaction run. Current cached pfn points the pageblock that
we already checked that it has no suitable pages. So, we uselessly
checks whether pageblock is valid for compaction and skip-bit is
set or not at next compaction run. This is really minor overhead, but,
it's better to remove any inefficiency.

This patch sets scanner's cached pfn to next pfn of checked pageblock
to discard this inefficiency.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>