power: correct some PageHighMem() usages by using is_highmem()
PageHighMem() actually represents page is direct mapped or not.
It is implemented by is_highmem_idx(page_zone(page)) since pages on
highmem zone would not be mapped directly. So, we uses them mixed.

But, there is a special zone, ZONE_MOVABLE. It is only usable for
non-kernel memory (GFP_HIGHUSER_MOVABLE is needed to allocated page
from it), so we need to count number of it's pages to highmem counter
even if PageHighMem() returns false.

So, this patch try to distinguish these two usages in power management
code. Change is simply mechanical. If PageHighMem() is used for creating
temporary mapping by using kmap_atomic(), leave it as is. Otherwise, it
would be used for checking number of usable pages and we should use
is_highmem() in this case.

Since PageHighMem() is implemented by is_highmem_idx(), there is no
functional change for now.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>