mm: implement PageHighMem() by using max_low_pfn rather than is_highmem()
PageHighMem() is used to check whether the page is direct mapped or not.
And, then, we create temporary mapping with the page if it isn't
direct mapped page. is_highmem() and is_highmem_idx() would be introduced
to represent the zone is highmem or not, but, nowadays, it is also used
to check whether the zone is usable for kernel memory or not. So, we need
to distinguish these usages.

This patch implement PageHighMem() by using max_low_pfn rather than
is_highmem() since is_highmem() will be used just for different meaning
that the zone is un-usable for kernel memory. I think that
using max_low_pfn will give same semantic and there is no functional
change from this patch.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>