mm/slub: do not consider allocation as memory reserved one if it has a fallback
If slab allocation request with GFP_ATOMIC comes and it requires
another slab page, SLUB tries to allocate high-order pages with
GFP_ATOMIC flag. The problem in this situation is that it causes
fragmentation. High-order request with GFP_ATOMIC changes migratetype
of the pageblock which allocated page is on to prepare further high-order
atomic requests. And, it is turned back to random migratetype.
If this process is repeated, fragmentation spread out whole memory space.

TO BE CONTINUE

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>