vchecker: introduce the valid access checker
Vchecker is a dynamic memory error detector. It provides a new
debug feature that can find out an un-intended access to valid area.
Valid area here means the memory which is allocated and allowed
to be accessed. Most of debug feature focused on finding out
un-intended access to in-valid area, for example, out-of-bound
access and use-after-free, and, there are many good tools for it.
But, as far as I know, there is no good tool to find out un-intended
access to valid area and this kind of problem is really hard to detect.

Idea to implement this feature is so simple. Thanks to compile-time
instrumentation, we can audit all the accesses to memory. However,
since almost accesses to valid area are usually valid, we need a way
to distinguish an in-valid access. What this patch provides is
the interface to describe the information about the address that
user are interested on and the judgement criteria for in-valid access.
With this information, we can easily detect in-valid access to valid area.

For now, two kinds of criteria are supported. One is the value checker
that checks written value to the designated address. The other is the
callstack checker that checks in-valid callstack when read/write happens.
It is more than welcome if someone introduces more checkers.

Now, this debug feature is only permitted for slab object. Coverage
would be extended soon.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>