mm/kasan: mark/unmark the target range that is for original shadow memory
Now, we have the per-page shadow. The purpose of the per-page shadow is
to check the page that is just used/checked in page size granularity.
File cache pages/anonymous page are in this category. The other
category is for byte size granularity. Global variable, kernel stack
and slab memory are in this category.

This patch distinguishes them and mark the page that should be checked by
the original shadow. Check for this page will be performed by original
shadow so we don't lose any checking accuracy even if we check some
other pages by using per-page shadow.

Note that there is no code for global variable since it is a static area
and it will be directly handled by architecture specific code.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>