kasan: support to keep object layout as much as possible
KASAN is a object layout sensitive feature that assumes that similar
location based on the object will be broken again. If object layout
is changed, this behaviour would be changed easily and we cannot
detect the problem.

To help this situation, this patch introduces a boot param that disables
the feature to change object layout. With enabling this option, kmem_cache
will not be merged. KASAN cannot have proper redzone so it is possible
to miss some type of OOB. And, KASAN doesn't quarantine the object
so UAF would be missed more. Lastly, KASAN will not report alloc/free
callstack of the object.

There are many drawbacks but nevertheless I think that this option is
important and useful because the situation that the user decides
to use this option would be the case that KASAN cannot solve the problem
with default setting.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>