lib/kasan: use the invalid read instead of the invalid write
Writing invalid value to the real object in the test would make the system
broken. Until now, it didn't happen since there is enough padding due to
KASAN specific data but the following patch will introduce a boot param
to completely remove this padding so writing invalid value to the real
object should not be allowed even in the test.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>