vchecker: per object memory for vchecker
To prepare per object memory for vchecker, we need to change the layout
of the object when kmem_cache initialization. Add a hook to change the
layout and call it on kmem_cache creation.

And, this memory should be initialized when object is populated. Do it
with another hook.

This memory will be used in the following patch.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>