vchecker: disable to store value write's callstack if kasan_keep_object_layout
There is an another user to use per-object data, value checker.
Value checker uses per-object data for stroing value write's callstack.
Since it cannot know and check the value written on time, it checks
the value in next time and show previous value writer's callstack if
the check is failed.

Although it is just temporary implementation, it's better to support
disabling it if kasan_keep_object_layout is true.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>