mm/kasan: support to map zero shadow for memory allocated by early allocator
There is a performance problem when per-page shadow is used
for CONFIG_KASAN_INLINE. Reason is that memory allocated by
early allocator (memblock or (no)bootmem) is backed by black shadow.
In CONFIG_KASAN_INLINE, initial sanity check for these memory fails
due to this black shadow and fall back to per-page shadow check.
Although a check will be passed here but it results in function call
thus performance is hurt.

Solution is that mapping zero shadow for memory allocated by these
early allocator. It is the same as original KASAN strategy and
it fixes performance problem for per-page shadow approach.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>